refactor: remove post-build script
